requirements.txt
gunicorn==20.0.4
Pillow==8.1.0
PyAudio==0.2.11
SpeechRecognition==3.8.1
